moved execute_notebooks functionality into build_docs module.
